Moved progress helper.
