+++ src/io/flutter/actions/OpenInXcodeAction.java
@@ -21,7 +21,7 @@ import com.intellij.openapi.vfs.VirtualFile;
 import io.flutter.FlutterMessages;
 import io.flutter.FlutterUtils;
 import io.flutter.pub.PubRoot;
-import io.flutter.run.daemon.ProgressHelper;
+import io.flutter.utils.ProgressHelper;
 import io.flutter.sdk.FlutterSdk;
 import io.flutter.utils.FlutterModuleUtils;
 import org.jetbrains.annotations.NotNull;
